build: fix das_hash_map C4244, drop /wd4244 from strict warnings
rehash_into indexed eastl::vector::operator[] (size_type == unsigned int in
the EASTL config) with a uint64_t slot index, the lone C4244 source under
/W4 /WX. The hot paths (find_index/reserve_slot) already probe via raw
.data() pointers, so they never warned. Make the slot index size_t in all
four rehash_into variants (map/set, regular + insert-only): the residual
index narrowing is then C4267 — same as the existing hashes_[i] size_t
indexing — which stays suppressed.

With the C4244 gone, drop /wd4244 from DAS_APPLY_STRICT_WARNINGS.

+++ include/das_hash_map/das_hash_map.h
@@ -538,7 +538,7 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh <= daslang_hash_map_detail::HASH_KILLED ) continue;
-            uint64_t j = kh & mask;
+            size_t j = kh & mask;
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
                 j = daslang_hash_map_detail::next_probe_slot(j, mask);
             }
@@ -851,7 +851,7 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh == daslang_hash_map_detail::HASH_EMPTY ) continue;
-            uint64_t j = kh & mask;
+            size_t j = kh & mask;
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
                 j = daslang_hash_map_detail::next_probe_slot(j, mask);
             }
@@ -1074,7 +1074,7 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh <= daslang_hash_map_detail::HASH_KILLED ) continue;
-            uint64_t j = kh & mask;
+            size_t j = kh & mask;
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
                 j = daslang_hash_map_detail::next_probe_slot(j, mask);
             }
@@ -1265,7 +1265,7 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh == daslang_hash_map_detail::HASH_EMPTY ) continue;
-            uint64_t j = kh & mask;
+            size_t j = kh & mask;
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
                 j = daslang_hash_map_detail::next_probe_slot(j, mask);
             }
